Unit_3/Airflow/requirements.txt
beautifulsoup4==4.11.1; python_full_version >= "3.6.0"
certifi==2022.9.24; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
colorama==0.4.5; python_version >= "2.7" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and platform_system == "Windows"
contourpy==1.0.5; python_version >= "3.8"
cycler==0.11.0; python_version >= "3.8"
cython==0.29.32; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
et-xmlfile==1.1.0; python_version >= "3.6"
fonttools==4.38.0; python_version >= "3.8"
idna==3.4; python_version >= "3.7" and python_version < "4"
joblib==1.2.0; python_version >= "3.8"
kiwisolver==1.4.4; python_version >= "3.8"
lxml==4.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
matplotlib==3.6.0; python_version >= "3.8"
missingno==0.5.1
mplcyberpunk==0.6.0
mpmath==1.2.1; python_version >= "3.8"
numpy==1.23.4; python_version >= "3.8"
openpyxl==3.0.10; python_version >= "3.6"
packaging==21.3; python_version >= "3.8"
pandas==1.5.1; python_version >= "3.8"
pathlib==1.0.1
patsy==0.5.3; python_version >= "3.7"
pillow==9.2.0; python_version >= "3.8"
plotly==5.10.0; python_version >= "3.6"
pmdarima==2.0.1; python_version >= "3.7"
psycopg2-binary==2.9.5; python_version >= "3.6"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.8"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
pytz==2022.5; python_version >= "3.8"
pyyaml==6.0; python_version >= "3.6"
qbstyles==0.1.4; python_version >= "3.5"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
scikit-learn==1.1.2; python_version >= "3.8"
scipy==1.9.3; python_version >= "3.8"
seaborn==0.12.1; python_version >= "3.7"
setuptools-scm==7.0.5; python_version >= "3.8"
shutup==0.2.0; python_version >= "3.5" and python_version < "4.0"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
soupsieve==2.3.2.post1; python_version >= "3.6" and python_full_version >= "3.6.0"
statsmodels==0.13.2; python_version >= "3.7"
sympy==1.11.1; python_version >= "3.8"
tenacity==8.1.0; python_version >= "3.6"
threadpoolctl==3.1.0; python_version >= "3.8"
tomli==2.0.1; python_version >= "3.8"
tqdm==4.64.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
typing-extensions==4.4.0; python_version >= "3.8"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
wordcloud==1.8.2.2
